lab4_sql/requirements.txt
Django>=4.0,<5.0
psycopg2>=2.9
python-decouple==3.8
djangorestframework==3.14.0
django-filter==24.3
pillow==11.0.0
python-magic==0.4.27